Add jwt token to cookie (Set-Cookie)
