+++ src/main/java/com/developerscorner/provider/security/JwtAuthenticationFilter.java
@@ -2,9 +2,11 @@ package com.developerscorner.provider.security;
 
 import javax.servlet.FilterChain;
 import javax.servlet.ServletException;
+import javax.servlet.http.Cookie;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
+import java.util.Arrays;
 
 import org.springframework.lang.NonNull;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
@@ -31,15 +33,23 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
 	protected void doFilterInternal(@NonNull HttpServletRequest request,
 			@NonNull HttpServletResponse response, @NonNull FilterChain filterChain)
 			throws ServletException, IOException {
-		final String authHeader = request.getHeader("Authorization");
-		final String jwt;
+		Cookie[] cookies = request.getCookies();
+		String jwt = null;
+		if (cookies != null) {
+			System.out.println("-------------------------------------" + cookies.toString());
+			jwt = Arrays.stream(cookies)
+					.filter(cookie -> "jwt".equals(cookie.getName()))
+					.map(Cookie::getValue)
+					.findAny().orElse(null);
+		}
+
 		final String userEmail;
-		if (authHeader == null || !authHeader.startsWith("Bearer ")) {
-			System.out.println("auth header is null====== token ============" + authHeader);
+		System.out.println("auth header is null====== token ============" + jwt);
+		if (jwt == null) {
 			filterChain.doFilter(request, response);
 			return;
 		}
-		jwt = authHeader.substring(7).replace("\"", "");
+
 		userEmail = jwtService.extractUsername(jwt);
 		System.out.println("userrrrrrrrrrrr====== token ============" + jwt);
 		if (userEmail != null && SecurityContextHolder.getContext().getAuthentication() == null) {
@@ -52,7 +62,6 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
 				authToken.setDetails(
 						new WebAuthenticationDetailsSource().buildDetails(request));
 				SecurityContextHolder.getContext().setAuthentication(authToken);
-				response.setHeader("Authorization", jwt);
 			}
 		}
 		filterChain.doFilter(request, response);
